cleanup: deprecate helper.js
the helper.js was useful 2 years ago, but time has passed.

we have https://github.com/bitfinexcom/bfx-svc-test-helper now,
which abstracts much of the boilerplate needed to use the
helper in a single npm package.

another problem is that mocha loads all files in a directory,
in some projects this leads to `grenache-grape` installed as
dependency, even when the use the helper is not used, just
that mocha can run.

the third problem is that people started to patch their local
versions, without sending patches to upstream, which leads to
problems with upgrades.

this removes `helper.js` from the core skeleton. old projects that
still depend on it, can merge upstream and then keep it locally.